requirement.txt
# django
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-simplejwt
pillow
django-rest-swagger

